webscrapping/app/src/requirements.txt
flask
selenium
pandas
beautifulsoup4
sqlalchemy
python-dotenv
psycopg2-binary

